Move refresh into auth, try to address issue 3
This simplifies the logic around session refreshes, and also causes the session key to be stored in the global store after a session refresh (which was a bug - it should have been before too).  Hopefully this will address 3 but I cannot test it as a dev channel and I'm not sure I have a good test for it anyway.